chore(password-field): move to packages
