+++ packages/components/fields/password-field/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './password-field';
+
+export { default as version } from './version';
