+++ packages/components/fields/password-field/src/password-field.form.story.js
@@ -6,11 +6,11 @@ import { action } from '@storybook/addon-actions';
 import { withKnobs, select } from '@storybook/addon-knobs/react';
 import Spacings from '@commercetools-uikit/spacings';
 import { PrimaryButton, SecondaryButton } from '@commercetools-uikit/buttons';
-import Section from '../../../../.storybook/decorators/section';
-import FormikBox from '../../../../.storybook/decorators/formik-box';
-import Readme from './README.md';
+import Section from '../../../../../.storybook/decorators/section';
+import FormikBox from '../../../../../.storybook/decorators/formik-box';
+import Readme from '../README.md';
 import PasswordField from './password-field';
-import PasswordInput from '../../inputs/password-input';
+import PasswordInput from '../../../inputs/password-input';
 
 // Cool stuff to try in this story:
 //  - Click the "Password" label and see how the input is focused automatically
