--- src/components/fields/password-field/index.js
@@ -1,1 +1,0 @@
-export { default } from './password-field';
